refactor: replace utility macros with guard functions
- add ThrowIfFalse to Guards
- remove macro-based validation from Utils
- simplify error handling with reusable guard functions

+++ EngineSFMLBackEnd/Code/UI/SFMenuCursor.cpp
@@ -28,7 +28,7 @@ void SFMenuCursor::SetScale(const Vector2f& cellSize)
     auto texSize = spr->GetTextureSize();
 
     // compare against local (texture) size, not global GetSize()
-    Vector2f tex = Vector2f(static_cast<int>(texSize.x), static_cast<int>(texSize.y));
+    Vector2f tex = Vector2f(static_cast<float>(texSize.x), static_cast<float>(texSize.y));
     Vector2f scaleXY = { cellSize.x / tex.x, cellSize.y / tex.y };
 
     float scale = std::min(scaleXY.x, scaleXY.y);
